优化文件写入、写出逻辑;优化配置文件相关;完成XPTask;ClassUtil新增缓存

+++ src/main/java/me/xpyex/plugin/xplib/bukkit/util/files/FileUtil.java
@@ -1,36 +1,29 @@
 package me.xpyex.plugin.xplib.bukkit.util.files;
 
+import java.io.BufferedWriter;
 import java.io.File;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.util.Scanner;
+import java.nio.charset.StandardCharsets;
+import java.nio.file.Files;
+import java.util.List;
 import me.xpyex.plugin.xplib.bukkit.util.RootUtil;
 import me.xpyex.plugin.xplib.bukkit.util.value.ValueUtil;
 
 public class FileUtil extends RootUtil {
+    /**
+     * 创建文件
+     *
+     * @param target   目标文件
+     * @param replaced 如果所在位置已存在文件，是否覆盖
+     * @throws IOException 文件异常
+     */
     public static void createNewFile(File target, boolean replaced) throws IOException {
-        if (replaced) {
+        ValueUtil.notNull("File不应为null", target);
+        if (replaced && target.exists()) {
             target.delete();
         }
-        while (!target.exists()) {
-            Throwable ex = null;
-            try {
-                target.createNewFile();
-            } catch (Throwable e) {
-                ex = e;  //当没权限，或者文件所在的目录还未创建
-            }
-            File parent = target;  //先定义文件自身
-            while (ex != null) {  //此时文件尚未被创建，故尝试创建父文件夹
-                ValueUtil.mustTrue("无法创建文件 " + target.getPath() + " ,原因可能是Java没有访问权限", parent != null);
-                parent = parent.getParentFile();  //寻找所在目录
-                try {
-                    parent.mkdirs();  //尝试创建所在目录及其所有父目录
-                    ex = null;  //打破第二个while
-                } catch (Throwable e) {
-                    ex = e;  //重启第二个while
-                }
-            }
-        }
+        target.getParentFile().mkdirs();
+        target.createNewFile();
     }
 
     /**
@@ -41,12 +34,11 @@ public class FileUtil extends RootUtil {
      * @throws IOException 文件异常
      */
     public static String readFile(File target) throws IOException {
-        Scanner in = new Scanner(target, "UTF-8");
-        StringBuilder builder = new StringBuilder();
-        while (in.hasNextLine()) {
-            builder.append(in.nextLine()).append("\n");
+        List<String> content = Files.readAllLines(target.toPath(), StandardCharsets.UTF_8);
+        if (content.isEmpty()) {
+            return "";
         }
-        return builder.toString();
+        return String.join("\n", content);
     }
 
     /**
@@ -54,21 +46,20 @@ public class FileUtil extends RootUtil {
      *
      * @param target  目标文本
      * @param content 要写出的内容
-     * @param attend  是否在原文本的内容基础续写新文本，否则覆写整个文件
+     * @param append  是否在原文本的内容基础续写新文本，否则覆写整个文件
      * @throws IOException 文件异常
      */
-    public static void writeFile(File target, String content, boolean attend) throws IOException {
+    public static void writeFile(File target, String content, boolean append) throws IOException {
         if (!target.exists()) {
             createNewFile(target, false);
         }
-        PrintWriter out = new PrintWriter(target, "UTF-8");
-        if (attend) {
-            out.println(content);
-        } else {
-            out.write(content);
+        try (BufferedWriter out = Files.newBufferedWriter(target.toPath(), StandardCharsets.UTF_8)) {
+            if (append) {
+                out.append(content);
+            } else {
+                out.write(content);
+            }
         }
-        out.flush();
-        out.close();
     }
 
     /**
